feat: Add Checkov scan step in GitHub workflows
Added a new step in the GitHub workflows for both plan and sync-and-deploy. The new step runs Checkov scan using the bridgecrewio/checkov-action@v12 action. The scan is performed on the current directory. This step ensures that the Terraform code is secure and compliant with best practices.

Updated the following files:
- .github/workflows/plan.yaml
- .github/workflows/sync-and-deploy.yaml

Also, added new documentation for Checkov in the README.MD file.

Affected files:
- _DOCUMENTATION/README.MD
- terraform/kyverno.tf
- terraform/policies/allowed-replica-limit.yaml
- terraform/policies/require-resource-limits.yaml